Move construction of node map tree into LiveObject class
There was a TODO for this from a while ago, but this wasn't done yet.
